dbarts variable-dropping stopgap; more warnings and errors
Fix variable-drop problem as a holdover (very overdue, I'm sorry this has taken so long - also not a permanent solution). This also removes a step that drops rows with NA's from the variable importance script, which I think is a holdover from a previous version of dbarts and known issues with it.